+++ src/Controller/File/GetElementFileController.php
@@ -4,20 +4,31 @@ declare(strict_types=1);
 
 namespace App\Controller\File;
 
-use App\Factory\Exception\Server501NotImplementedExceptionFactory;
+use App\Factory\Exception\Client400BadContentExceptionFactory;
+use App\Factory\Exception\Client400MissingPropertyExceptionFactory;
+use App\Factory\Exception\Client404NotFoundExceptionFactory;
 use App\Helper\Regex;
+use App\Security\AccessChecker;
+use App\Security\AuthProvider;
+use App\Service\ElementManager;
+use App\Type\AccessType;
+use Ramsey\Uuid\Rfc4122\UuidV4;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 
-/**
- * @SuppressWarnings("PHPMD.UnusedFormalParameter")
- */
 class GetElementFileController extends AbstractController
 {
     public function __construct(
-        private Server501NotImplementedExceptionFactory $server501NotImplementedExceptionFactory,
+        private AuthProvider $authProvider,
+        private AccessChecker $accessChecker,
+        private ElementManager $elementManager,
+        private UrlGeneratorInterface $router,
+        private Client400MissingPropertyExceptionFactory $client400MissingPropertyExceptionFactory,
+        private Client400BadContentExceptionFactory $client400BadContentExceptionFactory,
+        private Client404NotFoundExceptionFactory $client404NotFoundExceptionFactory,
     ) {
     }
 
@@ -31,6 +42,11 @@ class GetElementFileController extends AbstractController
     )]
     public function getElementFile(string $id, Request $request): Response
     {
-        throw $this->server501NotImplementedExceptionFactory->createFromTemplate();
+        $id = UuidV4::fromString($id);
+        $userId = $this->authProvider->getUserId();
+
+        if (!$this->accessChecker->hasAccessToElement($userId, $id, AccessType::READ)) {
+            throw $this->client404NotFoundExceptionFactory->createFromTemplate();
+        }
     }
 }