+++ src/Attribute/EndpointImplementsTusIo.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Attribute;
+
+use Attribute;
+
+#[Attribute]
+readonly class EndpointImplementsTusIo
+{
+    public function __construct()
+    {
+    }
+}