+++ src/Command/BackupLoadCommand.php
@@ -7,16 +7,23 @@ namespace App\Command;
 use App\DependencyInjection\DeactivatableTraceableEventDispatcher;
 use App\EventSystem\EntityManager\Event\ElementUpdateAfterBackupLoadEvent;
 use App\Factory\Exception\Server500LogicExceptionFactory;
+use App\Helper\Regex;
 use App\Service\AppStateService;
 use App\Service\ElementManager;
 use App\Service\RawToElementService;
+use App\Service\StorageUtilService;
 use App\Style\EmberNexusStyle;
 use App\Type\AppStateType;
+use AsyncAws\S3\Input\PutObjectRequest;
+use AsyncAws\S3\S3Client;
+use EmberNexusBundle\Service\EmberNexusConfiguration;
 use Laudis\Neo4j\Databags\Statement;
 use League\Flysystem\FilesystemOperator;
 use LogicException;
 use Predis\Client;
 use Ramsey\Uuid\Rfc4122\UuidV4;
+use Ramsey\Uuid\Uuid;
+use Ramsey\Uuid\UuidInterface;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
@@ -46,6 +53,9 @@ class BackupLoadCommand extends Command
     public function __construct(
         private ElementManager $elementManager,
         private CypherEntityManager $cypherEntityManager,
+        private EmberNexusConfiguration $emberNexusConfiguration,
+        private S3Client $s3Client,
+        private StorageUtilService $storageUtilService,
         private Client $redisClient,
         private FilesystemOperator $backupStorage,
         private RawToElementService $rawToElementService,
@@ -163,13 +173,69 @@ class BackupLoadCommand extends Command
         ));
     }
 
+    private function parseFilenameAsUuidFromPath(string $path): false|UuidInterface
+    {
+        $filename = basename($path);
+        $parts = explode('.', $filename, 2);
+        if (2 !== count($parts)) {
+            return false;
+        }
+        $name = $parts[0];
+        if (!preg_match(Regex::UUID_V4, $name)) {
+            return false;
+        }
+
+        return Uuid::fromString($name);
+    }
+
     private function loadFiles(): void
     {
         $this->io->startSection('Step 3 of 4: Loading Files');
-        $this->io->writeln('Currently not implemented.');
+        $progressBar = $this->io->createProgressBarInInteractiveTerminal($this->fileCount);
+        $progressBar?->display();
+        $files = $this->backupStorage->listContents($this->backupName.'/file/', true);
+        $pageCount = 0;
+        $totalCount = 0;
+        foreach ($files as $file) {
+            if (!$file->isFile()) {
+                continue;
+            }
+            $path = $file->path();
+            $fileId = $this->parseFilenameAsUuidFromPath($path);
+            if (false === $fileId) {
+                continue;
+            }
+            $element = $this->elementManager->getElement($fileId);
+            if (null === $element) {
+                $this->io->warning(sprintf(
+                    'Found file in backup without corresponding element; can not import file: %s',
+                    $path
+                ));
+                continue;
+            }
+
+            // todo: optimize upload for larger files using multipart-upload?
+            $resource = $this->backupStorage->readStream($path);
+            $this->s3Client->putObject(new PutObjectRequest([
+                'Bucket' => $this->emberNexusConfiguration->getFileS3StorageBucket(),
+                'Key' => $this->storageUtilService->getStorageBucketKey($fileId),
+                'Body' => $resource,
+            ]))->resolve();
+
+            ++$pageCount;
+            if ($pageCount >= $this->pageSize) {
+                $progressBar?->advance($pageCount);
+                $totalCount += $pageCount;
+                $pageCount = 0;
+            }
+        }
+        $this->elementManager->flush();
+        $progressBar?->advance($pageCount);
+        $progressBar?->clear();
+        $totalCount += $pageCount;
         $this->io->stopSection(sprintf(
             'Loaded <info>%d</info> files.',
-            0
+            $totalCount
         ));
     }