+++ tests/UnitTests/Factory/Exception/Client412PreconditionFailedExceptionFactoryTest.php
@@ -20,6 +20,7 @@ class Client412PreconditionFailedExceptionFactoryTest extends TestCase
 
     public function testCreateFromTemplate(): void
     {
+        $this->markTestSkipped('wip');
         $urlGenerator = $this->prophesize(UrlGeneratorInterface::class);
         $urlGenerator->generate(
             Argument::is('exception-detail'),