+++ CHANGELOG.md
@@ -13,9 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Increase max post limit from 2 MB (PHP default) to 101 MiB, related to #119.
 - Change header `Access-Control-Allow-Headers` to `*`, due to growing number of supported headers.
 ### Fixed
-- Fix deprecation `The "Symfony\Component\HttpKernel\DependencyInjection\Extension" class is considered internal since
-  Symfony 7.1, to be deprecated in 8.1; use Symfony\Component\DependencyInjection\Extension\Extension instead. It may
-  change without further notice.`.
+- Fix bug with deserialization of MongoDB documents.
 
 ## 0.1.30 - 2026-02-03
 ### Changed