+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add S3 status check to healthcheck command, closes #200.
 - Add support to delete files in S3 / object storage through the `php bin/console database:drop` command, closes #309.
+- Add support for `Content-Disposition` HTTP header, including transliteration of non-latin-characters, closes #447.
 ### Changed
 - Upgrade PHP to 8.5.2, closes #448.
 - Increase max post limit from 2 MB (PHP default) to 101 MiB, related to #119.