+++ CHANGELOG.md
@@ -5,11 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## 0.1.31 - 2026-02-10
 ### Changed
-- Upgrade PHP to 8.5.2, closes #448.
+- Upgrade PHP to 8.5.2, closes [#448].
 
 ### Fixed
-- Fix bad release 0.1.30, by fixing the CI, see #454.
+- Fix bad release 0.1.30, by fixing the CI, see [#454].
 
 ## 0.1.30 - 2026-02-03
 ### Changed